+++ backend/src/daoBean/AnalitikaMagacinskeKarticeDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.AnalitikaMagacinskeKartice;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface AnalitikaMagacinskeKarticeDaoLocal extends GenericDao<AnalitikaMagacinskeKartice, Long>{
+public interface AnalitikaMagacinskeKarticeDaoLocal extends GenericDao<AnalitikaMagacinskeKartice, Integer>{
 
 }
