+++ backend/src/daoBean/GrupaArtikalaDao.java
@@ -8,6 +8,6 @@ import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDaoBean;
 
 @Stateless
 @Local(GrupaArtikalaDaoLocal.class)
-public class GrupaArtikalaDao extends GenericDaoBean<GrupaArtikala, Long> implements GrupaArtikalaDaoLocal{
+public class GrupaArtikalaDao extends GenericDaoBean<GrupaArtikala, Integer> implements GrupaArtikalaDaoLocal{
 
 }
