+++ backend/src/daoBean/GrupaArtikalaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.GrupaArtikala;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface GrupaArtikalaDaoLocal extends GenericDao<GrupaArtikala, Long>{
+public interface GrupaArtikalaDaoLocal extends GenericDao<GrupaArtikala, Integer>{
 
 }
