+++ backend/src/daoBean/JedinicaMereDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.JedinicaMere;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface JedinicaMereDaoLocal extends GenericDao<JedinicaMere, Long> {
+public interface JedinicaMereDaoLocal extends GenericDao<JedinicaMere, Integer> {
 
 }
