+++ backend/src/daoBean/MagacinDaoLocal.java
@@ -3,7 +3,7 @@ package daoBean;
 import model.Magacin;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface MagacinDaoLocal extends GenericDao<Magacin, Long>{
+public interface MagacinDaoLocal extends GenericDao<Magacin, Integer>{
