+++ backend/src/daoBean/PopisnaKomisijaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.PopisnaKomisija;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface PopisnaKomisijaDaoLocal extends GenericDao<PopisnaKomisija, Long> {
+public interface PopisnaKomisijaDaoLocal extends GenericDao<PopisnaKomisija, Integer> {
 
 }
