+++ backend/src/daoBean/PopisniDokumentDao.java
@@ -8,6 +8,6 @@ import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDaoBean;
 
 @Stateless
 @Local(PopisniDokumentDaoLocal.class)
-public class PopisniDokumentDao extends GenericDaoBean<PopisniDokument, Long> implements PopisniDokumentDaoLocal{
+public class PopisniDokumentDao extends GenericDaoBean<PopisniDokument, Integer> implements PopisniDokumentDaoLocal{
 
 }
