+++ backend/src/daoBean/PopisniDokumentDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.PopisniDokument;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface PopisniDokumentDaoLocal extends GenericDao<PopisniDokument, Long>{
+public interface PopisniDokumentDaoLocal extends GenericDao<PopisniDokument, Integer>{
 
 }
