+++ backend/src/daoBean/StavkaPopisaDao.java
@@ -8,6 +8,6 @@ import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDaoBean;
 
 @Stateless
 @Local(StavkaPopisaDaoLocal.class)
-public class StavkaPopisaDao extends GenericDaoBean<StavkaPopisa, Long> implements StavkaPopisaDaoLocal{
+public class StavkaPopisaDao extends GenericDaoBean<StavkaPopisa, Integer> implements StavkaPopisaDaoLocal{
 
 }
