+++ backend/src/daoBean/StavkaPopisaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.StavkaPopisa;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface StavkaPopisaDaoLocal extends GenericDao<StavkaPopisa, Long> {
+public interface StavkaPopisaDaoLocal extends GenericDao<StavkaPopisa, Integer> {
 
 }
