+++ backend/src/daoBean/StavkaPrometnogDokumentaDaoLocal.java
@@ -3,6 +3,6 @@ package daoBean;
 import model.StavkaPrometnogDokumenta;
 import rs.ac.uns.ftn.xws.sessionbeans.common.GenericDao;
 
-public interface StavkaPrometnogDokumentaDaoLocal extends GenericDao<StavkaPrometnogDokumenta, Long>{
+public interface StavkaPrometnogDokumentaDaoLocal extends GenericDao<StavkaPrometnogDokumenta, Integer>{
 
 }
